fixed some type

+++ tasks/pylaeva_s_max_elem_matrix/tests/functional/main.cpp
@@ -37,7 +37,7 @@ class PylaevaSMaxElemMatrixFuncTests : public ppc::util::BaseRunFuncTests<InType
     }
     
     size_t size = 0;
-    std::vector<double> input;
+    std::vector<int> input;
     input.reserve(size);
     int max;
